Completed modifications to applications and host resource scheduling. Testing not yet done, work in progress.

+++ src/edu/uwo/csd/dcsim/application/workload/RandomWorkload.java
@@ -34,7 +34,7 @@ public class RandomWorkload extends Workload {
 	}
 	
 	@Override
-	protected double getCurrentWorkLevel() {
+	protected int getCurrentWorkLevel() {
 		return workLevel;
 	}
